requirements.txt
numpy==1.23.5
pandas==1.4.1
pyserial==3.5
PyWavelets==1.4.1
scikit_learn==1.0.2
scipy==1.8.0
statsmodels==0.13.2
tqdm==4.62.3
